+++ eras/babbage/impl/testlib/Test/Cardano/Ledger/Babbage/TreeDiff.hs
@@ -32,7 +32,7 @@ deriving newtype instance ToExpr CoinPerByte
 instance ToExpr (PlutusScript (BabbageEra c))
 
 -- PlutusContext
-instance ToExpr (PlutusPurpose AsIndex era) => ToExpr (BabbageContextError era)
+instance ToExpr (PlutusPurpose AsIx era) => ToExpr (BabbageContextError era)
 
 -- PParams
 instance ToExpr (BabbagePParams StrictMaybe era)
@@ -67,7 +67,7 @@ instance
 instance
   ( Era era
   , ToExpr (PredicateFailure (EraRule "UTXO" era))
-  , ToExpr (PlutusPurpose AsIndex era)
+  , ToExpr (PlutusPurpose AsIx era)
   , ToExpr (PlutusPurpose AsItem era)
   , ToExpr (TxCert era)
   ) =>